HSEARCH-3103 Clarify what DoubleMultiValuesSource and LongMultiValuesSource do
